requirements.txt
PyPDF2
faiss-cpu
numpy
sentence-transformers
streamlit